Rewrote to use a single stack
Pushing and popping needs to happen on a single stack. This requires that a virtual AST be built up for all the elements, so I pulled it off, using Spring Data Neo4j. It seems to work perfectly with only a few superfluous issues appearing, caused by gaps in docbook that are more strictly prevents with asciidoctor.